Disable some global flags with WASM
