+++ pkg/config/config.go
@@ -44,16 +44,6 @@ type Config struct {
 
 var DefaultProcessTimeoutSec = 60 * 60 * 2 // 2時間
 
-// InitConfig 指定のFlagSetにConfigへ値を設定するためのフラグを登録する
-func InitConfig(flags *pflag.FlagSet) {
-	initCredentialConfig(flags)
-	initOutputConfig(flags)
-	initDebugConfig(flags)
-	// misc flags
-	flags.IntP("process-timeout-sec", "", 0, "number of seconds before the command execution is timed out")
-	flags.BoolP("version", "v", false, "show version info")
-}
-
 // LoadConfigValue 指定のフラグセットからフラグを読み取り*Flagsを組み立てて返す
 func LoadConfigValue(flags *pflag.FlagSet, errW io.Writer, skipLoadingProfile bool) (*Config, error) {
 	o := &Config{
@@ -67,23 +57,6 @@ func LoadConfigValue(flags *pflag.FlagSet, errW io.Writer, skipLoadingProfile bo
 
 	o.loadConfig(flags, errW)
 	return o, o.Validate(false)
-}
-
-func initCredentialConfig(fs *pflag.FlagSet) {
-	fs.StringP("profile", "", "", "the name of saved credentials")
-	fs.StringP("token", "", "", "the API token used when calling SAKURA Cloud API")
-	fs.StringP("secret", "", "", "the API secret used when calling SAKURA Cloud API")
-	fs.StringSliceP("zones", "", []string{}, "permitted zone names")
-}
-
-func initOutputConfig(fs *pflag.FlagSet) {
-	fs.BoolP("no-color", "", false, "disable ANSI color output")
-}
-
-func initDebugConfig(fs *pflag.FlagSet) {
-	fs.BoolP("trace", "", false, "enable trace logs for API calling")
-	fs.BoolP("fake", "", false, "enable fake API driver")
-	fs.StringP("fake-store", "", "", "path to file store used by the fake API driver")
 }
 
 func (o *Config) IsEmpty() bool {
